data-streaming-pipeline/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.13.0
google-cloud-storage==2.13.0
google-genai==0.3.0
google-cloud-secret-manager==2.16.0
